+++ src/communities/manager.ts
@@ -40,10 +40,22 @@ export default class CommunitiesManager {
     }
   }
 
-  public create = async (certs: CertsData): Promise<CommunityData> => {
+  public createNetwork = async () => {
+    const ports = await getPorts()
+    const hiddenService = await this.connectionsManager.tor.createNewHiddenService(ports.libp2pHiddenService, ports.libp2pHiddenService)
+    await this.connectionsManager.tor.destroyHiddenService(hiddenService.onionAddress)
+    const peerId = await PeerId.create()
+    return {
+      hiddenService,
+      peerId: peerId.toJSON(),
+    }
+  }
+
+  public create = async (certs: CertsData,): Promise<CommunityData> => {
     const ports = await getPorts()
     const hiddenService = await this.connectionsManager.tor.createNewHiddenService(ports.libp2pHiddenService, ports.libp2pHiddenService)
     const peerId = await PeerId.create()
+
     const localAddress = await this.initStorage(peerId, hiddenService.onionAddress, ports.libp2pHiddenService, [peerId.toB58String()], certs)
     log(`Created community, ${peerId.toB58String()}`)
     return {